feat(exception) : 예외 메시지 추가
