+++ src/main/java/lotto/exception/ErrorPrinter.java
@@ -1,7 +1,7 @@
 package lotto.exception;
 
 public class ErrorPrinter {
-    public static void errorPrint(ErrorMessage message){
+    public static void errorPrint(InputErrorMessage message){
         System.out.println(message.getMessage());
     }
 }
