+++ src/test/java/lotto/exception/ErrorPrinterTest.java
@@ -1,7 +1,5 @@
 package lotto.exception;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,6 +7,6 @@ class ErrorPrinterTest {
     @Test
     @DisplayName("에러 메시지 출력 확인")
     void errorPrintTest(){
-        ErrorPrinter.errorPrint(ErrorMessage.PURCHASE_PRICE_CANT_DIVIDE_CLEARLY);
+        ErrorPrinter.errorPrint(InputErrorMessage.PURCHASE_PRICE_CANT_BE_DIVIDED_CLEARLY);
     }
 }
